Initial import of Kafka::FFI bindings
This sketches in a lot (91 of 248 functions) of the librdkafka
interface. Enough that I'm able to write a basic consumer against what
is there and have it be useful. The overall goal is with this commit is
to get a base of the lower two levels of abstraction (Kafka::FFI and
supporting classes) in place. There is enough here that the library
could be useful as is but still needs a lot of work.