Add Config settings for callbacks
There are a lot of callbacks... and I have a feeling that whole opaque
thing is going to get tricky in Ruby with garbage collection. I've
documented the main callbacks but ran out of steam with the socket ones
since I'm not feeling confident with them until I'm able to test them
properly.