Store references to callbacks in the Config to avoid GC
The callbacks are handed off to FFI at which point, unless otherwise
referenced, they are candidates for garbage collection. This adds a hold
on to those blocks so they will not be garbage collected as long as the
Config itself has references. The slightly simplifies the problem as
users only need to maintain the config and not every callback.