Refactor DeliveryReport lookup for Ruby 2.7
The documentation proved correct and using pointers to Ruby objects IS
DANGEROUS. Ruby 2.7 introduced a compacting garbage collector that I
believe is causing the issue because we were using Object ID to cheese a
pointer to an Object which can now be moved in memory.

I actually like this implementation better and I'm pretty proud of it.
Essentially, we allocate a pointer that we use as a hash key for the
object we want to reference. It is a little tricky but it cleanly
separates the two memory domains. One challenge was being able to free
the pointer once we were done with it, but using it as a key lets us
keep a reference which can later be freed after the report callback.