Refactor DeliveryReport locking and clean up mutexs
There is a non-zero chance that this is subtly wrong and will cause
issues. I can expect it to raise NoMethodError if @mutex gets removed
before the @mutex.synchronize call is made.

I'm mostly making this change to allow cleaning up of the mutex and
condition variable from #inspect output. So if there ARE any issues it
can be reverted.