Refactor Client.new to inject Config into the Client
Main goal was that we want the Client to maintain a reference to the
Config it was passed to ensure anything referenced isn't garbage
collected. I also split the overly long initializer into two more clear
parts avoiding focused on their role. This reads a lot clearer to me and
rather than a recursive call to new through some FFI magic, we have two
methods that are simpler.