Refactor OpaquePointer to allocate and initialize
This lets us clean up new for the objects to not have to check for a
pointer, but they will need a initialize method with a single ptr
argument. I'm happy that the `new` methods can now be focused on just
calling the librdkafka constructors (which then go through from_native
to initialize the object).