Refactor subscribe and subscription to avoid TopicPartitionList
This simplifies the API and should make it easier to work with. It
removes two potential cases of confusion and extra memory management.
For subscribe, it takes a TopicPartitionList but only cares about the
topic name, so setting any other values doesn't have an effect.
Similarly it makes sense that subscription(s) now just returns the
topics since to other fields in the TopicPartitionList are valid.