Adds DescribeConfigs and AlterConfigs admin interfaces
This implements the bulk of the structures needed to call
DescribeConfigs and AlterConfigs. I've tested them locally but haven't
added them to the Admin client due to needing to figure out some memory
issues. Queue#poll takes a block that destroy the event causing the
results to become invalid. Getting this commit out there so I can take a
longer think on how to solve that problem.