Introduces Opaque type for passing Ruby objects as opaque pointers
Librdkafka uses opaque pointers to let the application pass around
context into various places like callbacks and configs. With Ruby 2.6
and below it was possible to use a Ruby object's memory address but it
was kind of hacky and no longer works as of 2.7 (see a1322ec1).

This implements a look up registry for those opaque pointers and should
be generally usable. I've added it to the generic producer for testing
and am liking the API of it so far. I'm worried if it will be
generalizable for other opaque pointer uses and worry that it could
easily lead to memory leaks.