Add operation timeout to with_topic spec helper
Periodically have some failed tests because of topics not being
available. I want to get away from any sleeps in the tests and instead
having things only move forward when operations complete. This is partly
a test to see if it fixes the tests and I plan to revise the API since
it is clunky.