Refactor and reorganize the Admin API. Add remaining API
This adds CreatePartitions, AlterConfigs, and DescribeConfigs. It also
fixes a memory issue by introducing a new Admin::Result type that can be
destroyed when done with the results (since they are tied to the event).
This has also required changing Queue#poll to allow returning an event
instead of only taking a block since we need a different way to control
the lifecycle of an event.

I've removed Admin::Client and moved the API bindings to Client so they
are available from any producer or consumer. The convenience API has
been moved to Kafka::Admin and I've expanded on it to allow for waiting
for the operations to propogate. This had led to some refactorings in
with_topic and some tests. Overall I think this is a major improvement.

Note: Admin::Kafka is currently leaking memory since there is no way to
destroy the events. Will probably introduce classes that allow for
exporting the C structs into Ruby so they can be managed by the garbage
collector.