Change Producer#produce's block parameter to be called asynchronously
Since the Producer overwrites any delivery callback on the config (which
is still a problem) there was no way for an application to get the
status of every delivery unless they wanted to `wait` on every result.
This makes it possible for a consumer to care about the report but
maintain async publishing.