Changes Kafka::FFI::Client#metadata to return Ruby managed object
I've found that, when retrieving metadata, it's really _REALLY_ easy to
forget that the returned object is still managed by librdkafka. This
pulls all of the metadata into a Ruby managed object and frees the one
returned from librdkafka so we can have it easily garbage collected.

I've gone with the approach rather than using an autopointer because
there would be some bad interactions should client hold onto a Metadata
struct after the Client was destroyed.