Upgrade librdkafka to 1.5.0. Move LIBRDKAFKA_* constants
Upgrades to the latest release of librdkafka. There are a few upgrade
considerations outlined in the release notes and some changes to
defaults that may impact latency but improve throughput.

I've also moved the LIBRDKAFKA_VERSION and LIBRDKAFKA_CHECKSUM from the
Kafka module to the extension. They are only used when building the
extension and it would be better to depend on Kafka::FFI.version to get
the current version of librdkafka.

1: https://github.com/edenhill/librdkafka/releases/tag/v1.5.0