Adds Kafka::FFI::Queue#consume
This makes it easier to consume directly from topic's partition
directly, allowing a thread per partition.

The high level consumer will aggregate multiple partition queues into a
single queue for calls to Consumer#consumer_poll. The issue is that
messages appear to be placed on that queue in batches which means a
block of messages for a single partition will be processed before the
next partition. For applications that depend on low latency processing
of each partition, this can cause delays on a per partition basis.