Kafka::FFI::Config.get with a callback config key returns the callback
This should make get for callbacks more useful as it makes it possible
to get the callbacks which wasn't possible before. Also, it better fits
the style of librdkafka which returns a pointer to the callback
function. Before this change calling get would return the memory address
as a string which wasn't useful.