Kafka::FFI::Event#messages no longer allows a block
Allowing a block comes from an early version of the API but we've moved
to just returning an array in the other contexts. This should make the
API more consistent. The change also fixes calling #messages without a
block as Message.from_native didn't exist. I've managed to come up with
a way to test `fetch` messages and will need to investigate if we can
test other Event types as well.