ext: Try to avoid warnings about chdir in chdir
When installing kafka master with Bundler I've been seeing warnings
about Dir.chdir being called inside another chdir. I'm not sure if this
is related but it is the one place we call it.

Effectively this changes the cleanup from being relative to be explicit
which, I think, is a good change. The goal is still that ext/Rakefile
tasks can safely be called from any directory.