admin: Refactor admin results into Events. Add 1.6.x APIs
Adds the new DeleteGroups, DeleteRecords, and DeleteConsumerGroupOffsets
APIs.

The Admin APIs are all implemented as async requests that are sent into
librdkafka and the responses eventually appear on a Queue as an Event.
For all of the current APIs the responses have a single result type that
they expose (e.g. TopicResult, GroupResult, etc...) and the previous
implementation was returning that one result type for all responses. I
didn't understand that very well when I first implemented it and this
new version more closely matches what is in librdkafka. The new design
allows for getting more context from the Event as makes it possible for
a response to include multiple results as different methods should it be
needed.