ffi: Refactor Kafka::FFI::Event into multiple subtypes
In librdkafka there is one Event struct type but it is functionally
split into multiple subclasses as most functions only work with certain
event types. While implementing the Admin API refactoring in af54aff9 it
became clear that the other Event types needed a similar treatment.

With the explosion of types I decided to implement a registry system on
Event that maps enum values to the class for the type. One goal was to
make adding new events / results easier in the future. The first version
used a case statement but meant adding a new event type would mean also
changing the Event class. Now the mapping of type to class is in the
implementing class and will be easier to expand on.