Adds option to set and retrieve Opaque set for Events
Opaque lets you pass application objects through librdkafka to tie
together asynchronous events. This adds the APIs to include them when
making Admin API requests and to retrieve them from the results. It also
adds it to OffsetCommitEvent but I'm not sure if we support setting it
at the moment.