actions: Build librdkafka in a single task and reuse in tests
Each test is spending 1.5 minutes downloading and building librdkafka.
Even with parallelization that's a lot of extra time wasted across all
those tasks. This _should_ make all of tests run faster and will make
librdkafka cacheable.

It shouldn't matter which version of ruby we use to compile librdkafka
since it doesn't link against Ruby (we bind to it using ffi).