kafka: Switch to Bitnami docker builds and update tested versions
Switched to the Bitnami docker images because they look to be more up to
date than the wurstmeister images (which the newest is 2.8.1). Makes it
hard to test newer versions without available images.

I've also removed Kafka 2.5 and older from the test matrix since they're
older and no longer receiving updates. Needs to be a cutoff at some
point.