Support testing against Redpanda
Redpanda is an alternative implementation that supports most of the
Kafka API. I'm personally interested in it as it may be easier to
operate than Kafka as it doesn't depend on Zookeeper (which Kafka won't
need eventually).

There are two test cases that error badly on Redpanda (one causes a
segfault).