Check topic creation errors and rewrite timeout code in with_topic
We weren't checking the response form create_topic which could have
embedded an error that would be useful for debugging or keep other code
from working correctly.

The timeout library does some shenanigans to interrupt control of
running code that _may_ be put librdkafka in a bad state. This could
cause either the code to never return or the timeout to happen late but
given that it is just in the tests I'm okay with it.

I'm making this change to try to fix a redpanda test that is
consistently failing on GitHub.