actions: Support Broker IDs of 1 or 1001
This is to fix the tests running on Kafka 3.2 and 3.3. I have not dug
into the underlying issue deeply but it appears that as of 3.2 Kafka
running with Zookeeper cannot use the traditional Broker ID of 1001.
KRaft (which Kafka 3.4 is using) appears to allow it. I was also unable
to standardize on a Broker ID of 1 due to older versions crashing
outright when setting it.

Rather than spend too much time on this I decided to allow either in the
few tests that check for Broker IDs.